hw lesson_4 + work on the lesson_5
