+++ commands/RbacController.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace app\commands;
+
+
+class RbacController
+{
+
+}
